requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
msal
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pydantic
keyring
requests